+++ src/SFA.DAS.Payments.ProviderPayments.Application/Services/IMonthEndEventHandlerService.cs
@@ -6,6 +6,7 @@ namespace SFA.DAS.Payments.ProviderPayments.Application.Services
 {
     public interface IMonthEndEventHandlerService
     {
-        Task<List<long>> GetMonthEndUkprns(int collectionYear, int collectionMonth, CancellationToken cancellationToken = default(CancellationToken));
+        Task<List<long>> GetMonthEndUkprns(short collectionYear, byte collectionMonth,
+            CancellationToken cancellationToken = default(CancellationToken));
     }
 }
