+++ src/SFA.DAS.Payments.ProviderPayments.Application/Services/MonthEndEventHandlerService.cs
@@ -14,7 +14,8 @@ namespace SFA.DAS.Payments.ProviderPayments.Application.Services
             this.providerPaymentsRepository = providerPaymentsRepository;
         }
 
-        public Task<List<long>> GetMonthEndUkprns(int collectionYear, int collectionMonth, CancellationToken cancellationToken = default(CancellationToken))
+        public Task<List<long>> GetMonthEndUkprns(short collectionYear, byte collectionMonth,
+            CancellationToken cancellationToken = default(CancellationToken))
         {
             return providerPaymentsRepository.GetMonthEndUkprns(collectionYear, collectionMonth, cancellationToken);
         }
