+++ src/SFA.DAS.Payments.ProviderPayments.Application/Services/MonthEndService.cs
@@ -16,11 +16,14 @@ namespace SFA.DAS.Payments.ProviderPayments.Application.Services
             this.providerPaymentsRepository = providerPaymentsRepository;
         }
 
-        public async Task<List<PaymentModel>> GetMonthEndPayments(int collectionYear, int collectionPeriod, long ukprn, CancellationToken cancellationToken = default(CancellationToken))
+        public async Task<List<PaymentModel>> GetMonthEndPayments(
+            short collectionYear, 
+            byte collectionPeriod,
+            long ukprn, 
+            CancellationToken cancellationToken = default(CancellationToken))
         {
             var payments = await providerPaymentsRepository.GetMonthEndPayments(collectionYear, collectionPeriod, ukprn, cancellationToken);
             return payments.ToList();
         }
-
     }
 }
